all goods OK

+++ include/battle/guest.h
@@ -63,7 +63,7 @@ public:
     virtual void* getInstance();
     virtual u8 unit_d0();
     virtual bool unit_178();
-    virtual u32 id() const;
+    virtual u16 id() const;
     virtual Msg name() const;
     virtual s32 level() const;
     virtual bool unit_208();
